requirements.txt
streamlit==1.28.0
youtube-transcript-api==0.6.1
openai==1.3.0
python-dotenv==1.0.0
